Fix blockerList.json url-filter RE syntax
"url-filter" is a regular expression, so '.' must be escaped!

Moreover this field should really be: "^https?://([^/]+\\.)?example\\.com/", but the leading (almost-)wildcard "[^/]+" leads to a performance hit, so I elected to leave it out. Performance really requires fully-expanded host names, which would in turn make the number of entries much higher!
Note we can't use "^https?://example\\.com/" as a workaround since this would not match sub-domains.
Hence "example\\.com/"; could lead to a few false positives if a (bad) host name is in the URL, but this is quite rare in practice.